nucleusflow::core::traits

Trait Shareable

pub trait Shareable:
    Sized
    + Send
    + Sync
    + Debug {
    // Required method
    fn into_shared(self) -> Arc<RwLock<Self>> ;
}

Trait for types that can be shared between multiple processors.

This trait provides a standard way to wrap types in thread-safe reference-counted containers for sharing between processors.

Required Methods

fn into_shared(self) -> Arc<RwLock<Self>>

Converts the type into a shareable form.

Returns

Returns an Arc<RwLock<Self>> containing the shareable value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".