nucleusflow::process
Function write_content

pub fn write_content<P: AsRef<Path>>(
    path: P,
    content: &str,
) -> Result<(), ProcessError>

Writes content to a file at the specified path.

Arguments

path - A reference to the path of the file to write to.

content - The content to write to the file.

Errors

Returns a ProcessError::WriteError if writing the file fails, or ProcessError::InvalidPath if the path is invalid.